fix(chexagent): use FP32 model dtype to match remote-code image loader
Day-2 v7 run logged {'status': 'ok', 'counts': {'chexagent-2-3b': 0}} —
model loaded but 500/500 generation attempts failed with:

    RuntimeError: Input type (torch.cuda.FloatTensor) and weight type
    (CUDABFloat16Type) should be the same

The CheXagent-2-3b remote-code modeling file materializes image tensors as
FP32 internally regardless of the model's weight dtype (the image loader
is invoked via tokenizer.from_list_format and bypasses the standard
pixel_values pipeline — no dtype coordination). Loading the model in
bfloat16 (or casting to bfloat16 post-load) therefore causes a hard
mismatch on every forward pass.

Fix: leave the model at its default FP32 dtype. At 3B params * 4 bytes =
~12 GB VRAM, well within the H100 80 GB envelope. BF16 is the preferred
perf option but it's incompatible with this specific trust_remote_code
implementation.

Same change applied to CheXagentV2Baseline (verifier-mode loader).

+++ v5/eval/baselines.py
@@ -353,8 +353,12 @@ class CheXagentV2Baseline(BaselineVerifier):
             kwargs["token"] = token
         self.device = torch.device(device) if not isinstance(device, torch.device) else device
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_id, **kwargs)
+        # Default dtype (FP32). See CheXagentV2Generator for rationale: the
+        # remote-code image loader emits FP32 tensors regardless of model
+        # dtype, so explicit bfloat16 casting produces a type-mismatch at
+        # inference.
         self.model = AutoModelForCausalLM.from_pretrained(
-            self.model_id, device_map="auto", torch_dtype=torch.bfloat16, **kwargs,
+            self.model_id, device_map="auto", **kwargs,
         ).eval()
 
     @torch.no_grad()
